refactor: getting proposal execution data from backend
