+++ client/src/components/vote/CancelModal.tsx
@@ -1,5 +1,6 @@
 import { Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
+import { ProposalExecutionData } from 'api/proposals'
 import GrayCloseButton from 'components/GrayCloseButton/GrayCloseButton'
 import { useIsMobile } from 'nft/hooks'
 import { useState } from 'react'
@@ -7,7 +8,7 @@ import { ArrowUpCircle, X } from 'react-feather'
 import styled, { useTheme } from 'styled-components/macro'
 
 import Circle from '../../assets/images/blue-loader.svg'
-import { ProposalExecutionData, useCancelCallback } from '../../state/governance/hooks'
+import { useCancelCallback } from '../../state/governance/hooks'
 import { CustomLightSpinner, ThemedText } from '../../theme'
 import { ExternalLink } from '../../theme'
 import { ExplorerDataType, getExplorerLink } from '../../utils/getExplorerLink'
